Introduce internal implementation
